Look for emacs and test.el a little harder.
If the EMACS environment variable is set, use it. If an Aquamacs
bundle is given, find the emacs inside automatically.

If test.el is not found in the emacs load path, including in the
directories given on the command line, look for a test.el relative to
the test script.